Fix bug in `incompressibleDarcy`
If velocity is prescribed at the inlet and/or outlet, the value of the
velocity in the cells adjacent to the boundary was wrong for unknown
reasons, while the values in all the other cells were good.

We had to change how the velocity field is recalculated at the end of
`correctPressure`. As a consequence the solver now reimplements
`correctPressure.C` from the solver `incompressibleFLuid` in addition to
`momenentumPredictor.C`. The change is only 1 line, we use
`fcv::reconstruct` in lieu of the `HbyA` and `rAtU` mechanism.

Signed-off-by: Emmanuel LE TRONG <emmanuel.le-trong@cnrs-orleans.fr>