Add the `darcy` solver (compressible flow)
Which is the compressible counterpart of `incompressibleFluid`.

It is based on the `isothermalFluid` solver, but because the latter has
two private members we need to alter, and those in turn call other
private members functions, we end up duplicating a lot of code.

At the moment it passes the same tests as the incompressible solver.

Signed-off-by: Emmanuel LE TRONG <emmanuel.le-trong@cnrs-orleans.fr>